Check /cmdline for programe name as well as /exe.
This is useful for when the program was called via a symlink. For example on
my system `mplayer` is an alias to `mplayer2`. Since I launch it with
`mplayer` if I wanted to check the progress using cv I would expect to use `cv
-c mplayer` but that currently doesn't work. The original command line
(assuming argv[0] hasn't been modified) has this information:

    $ readlink `which mplayer`
    /etc/alternatives/mplayer
    $ readlink -f `which mplayer`
    /usr/bin/mplayer2

    $ readlink /proc/`pidof mplayer`/exe
    /usr/bin/mplayer2
    $ cat /proc/`pidof mplayer`/cmdline | tr '\0' ' ' | cut -d' ' -f1
    mplayer

Currently I am checking for a full basename match on the executable but it may
be nice to add a -f,--full like pgrep does. Also -g,--glob (fnmatch) my be
useful for partial matches with -c.

This adds an more overhead for every unsuccessful /exe lookup though :(. If
this becomes an option there are a couple of options like: only do this extra
check if -c was passed (but then symlinks for cp (eg from busybox) wouldn't
work), hiding it behind an option, caching.