Add --full argument to match against the entire command line.
Matches the search pattern against the entire command line from
/proc/pid/cmdline instead of just the command name (argv[0]).
Since we are currently doing an exact match it is only really useful with eg
`cv -fgc \*pattern\*`. Hence implies `--glob`. If we do support substring
matches at some point it might be nice to make `-f` require `-c`.

Since this modifies the string that is copied to `pidinfo_list[i].name` it
could print the wrong process name in the info printout. Specifically since
the null characters from /proc/pid/cmdline are converted to spaces we lose the
distinction between arguments with spaces in them and argument delimiters.
Since we only care about the binary name here then this means that if your
binary has a space in it only the first bit of the name will be printed. This
is surely rare though, no processes in my PATH have spaces in the name.

Removed FNM_PATHNAME on the fnmatch(..., cmdline) call or else it wasn't
matching patterns like `myprog\*` against `myprog /target/dir`.

Also, with `--full` the comparison is no longer against the command basename,
in case people wan't to match against the path or `./thinger` or whatever.