Optimization: only go through /proc once.
Now instead of going through proc for each binary we are interested in go
through proc once and loop in the process names for each pid.

I wasn't actually seeing any performance issues so this is really a premature
optimisation...

The diff for `find_pids_by_binary_name` may look really big but that is only
because the indentation changed because of the new inner loop. The only thing
that changed in the inner loop is that I changed a couple of `break`s into
`goto`s to break out of the outer loop. If you don't like `goto`s you could
just add an additional `max_procs` check at the top of the outer loop.

Changed proc_names to be no longer null terminated just so it matches up with
the array generated from the `-c` argument.