Allow to filter proc/fd results by directory.
Implement an `lsof` alike `-D` argument. Eg

    cv -gc \* -D /tmp

will look through all processes to see if they have a file open in /tmp. The
fd selection for the actual printout is also filtered although it is possible
to not filter it there, I can't think of a reason why you would want that
though.

Changes to the two main functions to accept directory filter arguments and to
accept NULL/0 for some values. If you call `find_fd_for_pid` with NULL for the
output list it will just return a count of interesting fds.

TODO: Only works for absolute paths at the moment. Need to use realpath.