Add flag -f to display all files from the monitored processes
Sometimes you want to display all files read and written by a process,
perhaps because the process acts as a filter (e.g. ffmpeg), and
information about all files is useful. Flag `-f`, or `--full-display`,
is added for this purpose.

The number of files in the output is still limited statically.